[Potential reverse] Disable loading of Default Pubs
With the default-savestate in place, we no longer need to load the
default publications on intialization.

The order of publications in the Config becomes less important this way.

Of course, this makes adding a new default publication a bit more
difficult, but overall the process is simplified, including quick
drafting of changes.